Refactor index and definitions files to correctly export all types
